Create README.md
# Twitch Background Changer

**Status:** 🚧 In Development

This browser extension allows you to set a custom background image on specific areas of Twitch.  
Please note that this project is currently in development and may have bugs or incomplete features.

---

## Warning

Use this extension at your own risk. If you are not sure what you are doing, we recommend **not using** this extension.  
The developer is not responsible for any issues or damages caused by using this software.

---



## Installation

1. Clone or download this repository.  
2. Load it as an unpacked extension in your browser (Chrome/Edge)
3. Open Twitch and use the popup to upload your background image.

---



## License

This project is licensed under the MIT License.